requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
llama-index
llama-index-readers-file
llama-index-llms-openvino==0.2.2
llama-index-embeddings-openvino==0.2.1
llama-index-embeddings-huggingface==0.2.3

optimum-intel==1.20.0
transformers==4.45
diffusers==0.30.3
nncf==2.13.0
openvino==2024.4.1.dev20240926
openvino-genai==2024.4.1.0.dev20240926

torch==2.3.1
torchvision

librosa==0.10.2
onnx==1.16.1
qwen-vl-utils
Pillow
librosa
opencv-python
tqdm

gradio==4.44.1
modelscope
modelscope_studio

# jupyter
jupyterlab
ipywidgets
ipykernel>=5.0
ipython>=7.16.3
